[refactor] Modifications due to dir structure
