+++ src/main/java/mogakco/StudyManagement/scheduler/AbsentScheduleBatch.java
@@ -1,10 +1,12 @@
-package mogakco.StudyManagement.batch;
+package mogakco.StudyManagement.scheduler;
 
 import org.springframework.context.annotation.Configuration;
+
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+import mogakco.StudyManagement.service.common.LoggingService;
 import mogakco.StudyManagement.service.stat.StatService;
 
 @Configuration
@@ -12,18 +14,21 @@ import mogakco.StudyManagement.service.stat.StatService;
 @Service
 public class AbsentScheduleBatch {
     private final StatService statService;
+    private final LoggingService lo;
 
-    public AbsentScheduleBatch(StatService statService) {
+    public AbsentScheduleBatch(StatService statService, LoggingService lo) {
         this.statService = statService;
+        this.lo = lo;
     }
 
     @Scheduled(cron = "0 59 23 * * ?")
     public void executeDailyBatch() {
         try {
-            statService.createAbsentLog(null);
-            System.out.println("매일 23시 59분에 실행되는 배치 작업이 완료되었습니다.");
+            statService.createAbsentLog(lo);
+            System.out.println("일일 출석 로그가 저장되었습니다.");
         } catch (Exception e) {
-            System.err.println("배치 작업 도중 오류가 발생했습니다: " + e.getMessage());
+            System.err.println("출석 로그 저장을 위한 배치 작업 도중 오류가 발생했습니다: " + e.getMessage());
         }
     }
+
 }
