+++ src/test/java/mogakco/StudyManagement/controller/StatControllerTest.java
@@ -20,9 +20,9 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import mogakco.StudyManagement.batch.AbsentScheduleBatch;
 import mogakco.StudyManagement.dto.DTOReqCommon;
 import mogakco.StudyManagement.enums.LogType;
+import mogakco.StudyManagement.scheduler.AbsentScheduleBatch;
 import mogakco.StudyManagement.service.common.LoggingService;
 import mogakco.StudyManagement.service.stat.StatService;
 import mogakco.StudyManagement.util.DateUtil;
@@ -155,13 +155,15 @@ public class StatControllerTest {
     @DisplayName("부재 로그 배치 작업 성공")
     public void executeDailyBatch() {
         StatService statServiceMock = Mockito.mock(StatService.class);
+        LoggingService loggingServiceMock = Mockito.mock(LoggingService.class);
 
-        AbsentScheduleBatch absentScheduleBatch = new AbsentScheduleBatch(statServiceMock);
+        AbsentScheduleBatch absentScheduleBatch = new AbsentScheduleBatch(statServiceMock, loggingServiceMock);
 
         absentScheduleBatch.executeDailyBatch();
 
-        verify(statServiceMock, times(1)).createAbsentLog(null);
+        verify(statServiceMock, times(1)).createAbsentLog(loggingServiceMock);
     }
+
     ////////////////////////////////////////////////////////
 
     @Test
